BigTIFF upgrade in progress - widespread temp mess - does not compile now

+++ libtiff/tif_close.c
@@ -74,7 +74,7 @@ TIFFCleanup(TIFF* tif)
 	/* Clean up custom fields */
 	if (tif->tif_nfields > 0)
 	{
-		size_t  i;
+		uint32 i;
 
 	    for (i = 0; i < tif->tif_nfields; i++) 
 	    {
